# perf(api): use custom serialization for recipe list