--- a/backend/core/recipes/views.py
+++ b/backend/core/recipes/views.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 import logging
 from typing import Any, Iterable, Optional, cast