--- a/backend/core/recipes/views.py
+++ b/backend/core/recipes/views.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+
 import logging
 from typing import Any, Iterable, Optional, cast