This improves serialization from 1100ms to 125ms. Those numbers are best of three.

drf serialization
1088 ms
1101 ms
1229 ms

diy serialization
125 ms
175 ms
253 ms